feat: add model for Podcast
A model for Podcast has been added. This model is an early model, so it
may have small changes during the development. The idea is to use this
model along with Room to store Podcast subscribed by the user.

This commit adds a Podcast model class and its corresponding unitTest.

Close: #1