feat: initial MainActivity UI
This is the first step in order to build the MainActivity. The fragment
shown at the start of the application is PodcastListFragment, for this
reason this fragment is added too in this commit. Besides this, the
files needed to follow the MVVM model has been added.

This commit add the files needed as described above.

Relates: #11