feat: add initial recyclerview PodcastListFragment
To present and handle the different podcasts in PodcastListFragment, an
adapter whit its corresponding RecyclerView has been added. In addition
to this, some files has been modified for this purpose. This adapter is
still in its early version, it will be modified in future commits.

This commit adds the files needed to performed what is described above.

Relates: #11