feat: created VM and Repo for PodcastFragment
Now PodcastFragment uses a ViewModel and a Repository to obtain the
podcast and podcastEpisodes data. The PodcastEpisodeListAdapter needs
to be updated with the podcastEpisode information. All this information
is harcoded. This will change when the ivoox and Spotify APIs are added

This commit add the files and changes related to the work described.

Relates: #12