bug: fix PodcastListFragment crash
This fragment has two lists, one for names and another for images.
Besides this, another list contains the Podcast. The size of Podcast and
names/images was different, provoking a out of index error as we were
getting the index from the names/images and trying to use it in the
podcast list.

This commits removes the names/images list in favor of podcast.

Relates: #11