feat: add logic SearchFragment
Now the SearchFragment queries the Podcasts using the ViewModel
passing the query set by the user as parameter and updates the adapter
with the podcasts result. The Podcast data in PodcastRepository is
hardcoded until the APIs are implemented.

This commit add the changes in the necessary files to accomplish what
has been described above.

Relates: #14