feat: improve DownloadsFragment
Now this fragment queries the downloaded episodes from the repository
using the PodcastViewModel. This data is fake until the API are created.
Once the PodcastEpisodeFragment is created, clicking on one episode in
this fragment will open a PodcastEpisodeFragment to reproduce it.

Besides the activity_main layout has been improved.

This commit add the changes and files needed to accomplish the work
described before.

Relates: #22