feat: add option to select SwipeListener
The SwipeListener was set by default in all the instances of
PodcastEpisodeListAdapter. This is not how we want to proceed. It
doesn't make sense to have this listener in PodcastFragment as we are
not going to perform any action.

This commit adds the changes to use or not the SwipeListener in the
PodcastEpisodeListAdapter