feat: PodcastEpisodeFragment and remove navigation
An initial approach of the PodcastEpisodeFragment has been added. In
addition to this, the Navigation component is going to be remove as is
not an easy task to maintain the fragment state, so we are creating and
deleting fragments all the time.

This commit adds the files needed for PodcastEpisodeFragment and start
the purge of the Navigation component