feat: final decision navigation
After more and more testing and checking different apps such as Ivoox.
I decided to use the Navigation component to move around the app BUT
without saving fragment state.

Currently, is possible to save the instance of the fragments of the
BottonNavigationView. But not the rest. For this reason, I decided to
avoid saving any fragment instance at all.

The IMPORTANT thing about fragment state is how to handle the
PodcastEpsiodeFragment to guaranty a smooth use to the user.