feat: add AndroidSlidingUpPanel
Now the app uses AndroidSlidingUpPanel layout. Ivoox and Spotify use
a similar element to handle the audio reproduction allowing the user
to navigate between the different screens listening to the media.

To maintain the AndroidSlidingUpPanel independently of the screen
selected I followed this approach:

- MainActivity contains a Fragment(MainFragment)
 and BottonNavigationView
- MainFragment contains the AndroidSlidingUpPanel
- AndroidSlidingUpPanel contains a Fragment (this is changed by
the Navigation component) and the media reproducer (work in progress)

The idea is to hide the media reproducer and only show when need it.

Relates: #11