feat: add layout for MainFragment
The layout used by MainFragment shows an image of the episode, its
description and the ExoPlayer while the SlidingUpPanel's main layout
is shown. Otherwise (sliding layout), it shows an small image of the
episode, its name and a button to stop/start the ExoPlayer.

Currently the layout doesn't have any logic. It will be added in future
commits.

This commit adds the files and modifications required to achieve the
work defined above.

Relates: #13