refactor: first changes clean code
I have been reading the first chapters of Uncle Bob's Clean Code book.
Because of that, I'm refactoring the methods and comments according to
the explanations described in the book. This is the first commit
regarding this matter.

This commit add the changes to starting to follow the Clean Code
conventions