feat: add communication to update reproducer
The flow to update the MainFragment (reproducer) information is the following:

- When a episode is selected in PodcastFragment its updates the information on MainFragment.

To do so, I have added a shared ViewModel, ReproducerViewModel.

Besides this, the MainFragment.java class has been updated according to Clean Code.

Finally there are some small changes in the layouts of both fragments for testing purposes.

Relates: #11