feat: add swipe option
The option to swipe to delete elements of the screen has been added.

Now the slidingLayout containing the reproducer can be closed by swiping
to the right side of the screen.

The episodes in the DownloadsFragment have this option too. But only the
listener has been installed, no logic added.

Besides this, there have been some layouts changes to modularize them
and some small naming changes.

Relates: #11, #22