feat: create proper settings
Now the SettingsFragment shows the users that will be available to the
user. The logic of this settings needs to be coded.

The resources needed for the SettingsFragment have been created, new
fields in strings.xml and array.xml file. Besides this, some string in strings.xml have been renamed.

This commit adds the new files and changes for the SettingsFragment

Relates: #21