feat: finished podcast provider migration
Now all the podcast gathering info is performed by the Provider class.
All references to Ivoox/Spotify and DigitalPodcast have been removed.

Now the async task are run in parallel for better performance.

An ProgressDialog place holder has been added to. Is shown the first
time the user opens the apps. When the mayor part of parsing the podcast
information is done.

Relates: #18 #19