feat: add subscribe functionality
Now is possible to subscribe to podcasts.

The user has the option to select a podcast and click on the subscribe
button. Then the podcast will appear in the subscribe list.

The podcast subscribed to are stored in a database using Room. The unsubscribe part is not implemented yet, needs more work.