feat: avoid re-query database
Now we use LiveData as return type for getPodcast() PodcastDAO operation.
This means that we avoid re-querying the database to get new additions.

If a podcast is added, we will get notified by LiveData. Because of this,
we don't need the PodcastCallBack anymore.

This commit adds the changes described above.

Relates: #5 #13