refactor: Deleted MainFragment
Due to using a fragment has navhost instead of an activity, the back
button was not working properly. Indeed, this fragment was unnecessary
as all its logic could be implemented in MainActivity easily.

Besides this, some drawable resources have been renamed.

This commit adds the changes and work described above.

Relates: #11
Closes: #13