feat: add ads and improve PodcastListViewModel
Now the PodcastList and Search fragments shows ads to the user. The manifest information is missing as it's private.

Besides this, PodcastListViewModel has been modified to log a message
when there it cannot get episodes or podcast from the Repository. In addition, it returns an empty episode/podcast instead of null.