feat: add analytics
With the use of Google Analytics we get notified of the podcasts and
episodes selected by the users.

To do so, we get a reference to Analytics in PodcastListViewModel. To
ensure we are only getting one and separating this logic from the
fragments.

The size of the ad banner have changed.