feat: only reproduce audio
Now the ExoPlayer only reproduces the audio. In some cases, the podcast
contained an image, as all the podcasts don't contain an image to shown,
 I decided to not show the image at all.

Instead, all the podcasts will show the image associated to its episode
according to its url.

This commit adds the changes needed to achieve the work described above.