feat: add more exoplayer control and load screen
Now is possible to control the exoplayer by using the button of the
slidingPanel when it's minimized.

Besides, a loading indicator has been added. This indicator goes off
when the podcast data in PodcastListFragment is loaded correctly.

This commit add the changes and new files to achieve what is described
above.

Closes: #11, #12, #14 #15